// src/spec/threatgile.generated.ts
export interface Threagile {
// ...
  data_assets: {
    [k: string]: {
      /**
       * ID
       */
      id: string;
      /**
       * Description
       */
      description: string | null;
      /**
       * Usage
       */
      usage: "business" | "devops";
      /**
       * Tags
       */
      tags?: string[] | null;
      /**
       * Origin
       */
      origin?: string | null;
      /**
       * Owner
       */
      owner?: string | null;
      /**
       * Quantity
       */
      quantity: "very-few" | "few" | "many" | "very-many";
      /**
       * Confidentiality
       */
      confidentiality: "public" | "internal" | "restricted" | "confidential" | "strictly-confidential";
      /**
       * Integrity
       */
      integrity: "archive" | "operational" | "important" | "critical" | "mission-critical";
      /**
       * Availability
       */
      availability: "archive" | "operational" | "important" | "critical" | "mission-critical";
      /**
       * Justification of the rating
       */
      justification_cia_rating?: string | null;
      [k: string]: unknown;
    };
  };
// ...
}